fix(all): fixed some CI errors

+++ src/camera.cpp
@@ -14,9 +14,9 @@ void Camera::update(const gr::Render::Shader& shader, const gr::Window& window){
 
     glm::vec3 front;
     float yawAdjusted = rotation.y - 90.0f;
-    front.x = cos(glm::radians(yawAdjusted)) * cos(glm::radians(rotation.x));
-    front.y = sin(glm::radians(rotation.x));
-    front.z = sin(glm::radians(yawAdjusted)) * cos(glm::radians(rotation.x));
+    front.x = float(cos(glm::radians(yawAdjusted)) * cos(glm::radians(rotation.x)));
+    front.y = float(sin(glm::radians(rotation.x)));
+    front.z = float(sin(glm::radians(yawAdjusted)) * cos(glm::radians(rotation.x)));
     front = glm::normalize(front);
 
     glm::vec3 posGLM = {pos.x, pos.y, pos.z};
@@ -25,8 +25,8 @@ void Camera::update(const gr::Render::Shader& shader, const gr::Window& window){
 
     glm::mat4 view = glm::lookAt(posGLM, cameraTarget, up);
 
-    GLuint pLoc = glGetUniformLocation(shader.getProgram(), "uProjection");
-    GLuint vLoc = glGetUniformLocation(shader.getProgram(), "uView");
+    GLint pLoc = glGetUniformLocation(shader.getProgram(), "uProjection");
+    GLint vLoc = glGetUniformLocation(shader.getProgram(), "uView");
     glUniformMatrix4fv(pLoc, 1, GL_FALSE, &projection[0][0]);
     glUniformMatrix4fv(vLoc, 1, GL_FALSE, &view[0][0]);
 }
@@ -35,9 +35,9 @@ void Camera::moveForward(float speed){
     float yawRad = glm::radians(rotation.y - 90.0f);
 
     glm::vec3 forward;
-    forward.x = cos(yawRad);
+    forward.x = float(cos(yawRad));
     forward.y = 0.0f;
-    forward.z = sin(yawRad);
+    forward.z = float(sin(yawRad));
     forward = glm::normalize(forward);
 
     pos += forward * speed;
@@ -47,9 +47,9 @@ void Camera::moveBack(float speed){
     float yawRad = glm::radians(rotation.y - 90.0f);
 
     glm::vec3 forward;
-    forward.x = cos(yawRad);
+    forward.x = float(cos(yawRad));
     forward.y = 0.0f;
-    forward.z = sin(yawRad);
+    forward.z = float(sin(yawRad));
     forward = glm::normalize(forward);
 
     pos -= forward * speed;
@@ -59,9 +59,9 @@ void Camera::moveLeft(float speed){
     float yawRad = glm::radians(rotation.y - 90.0f);
 
     glm::vec3 forward;
-    forward.x = cos(yawRad);
+    forward.x = float(cos(yawRad));
     forward.y = 0.0f;
-    forward.z = sin(yawRad);
+    forward.z = float(sin(yawRad));
     forward = glm::normalize(forward);
 
     glm::vec3 right = glm::normalize(
@@ -75,9 +75,9 @@ void Camera::moveRight(float speed){
     float yawRad = glm::radians(rotation.y - 90.0f);
 
     glm::vec3 forward;
-    forward.x = cos(yawRad);
+    forward.x = float(cos(yawRad));
     forward.y = 0.0f;
-    forward.z = sin(yawRad);
+    forward.z = float(sin(yawRad));
     forward = glm::normalize(forward);
 
     glm::vec3 right = glm::normalize(
